Combine enum and struct module
These modules are tightly coupled. It makes sense for them to be in one types module.
